requirements.txt
python-dotenv = 0.21.0
requests = 2.28.1